day125/gcp-log-integration/requirements.txt
# GCP Integration Dependencies - Updated May 2025
google-cloud-logging==3.8.0
google-cloud-pubsub==2.20.0
google-auth==2.29.0
google-api-core==2.18.0

# Async and Processing
aiohttp==3.9.5

# Configuration and Data
pyyaml==6.0.1
pydantic==2.7.1
python-dotenv==1.0.1

# Dashboard and Monitoring
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Utilities
tenacity==8.3.0
structlog==24.1.0
colorama==0.4.6
